refactor CovidStats to CovidStatsDTO
