+++ Project_2/LocationStatusAPI/src/main/java/com/Project_2_Location_Status_API/Controllers/CovidApiController.java
@@ -1,7 +1,6 @@
 package com.Project_2_Location_Status_API.Controllers;
 
-import com.Project_2_Location_Status_API.Entities.CovidStats;
-import com.Project_2_Location_Status_API.Entities.Status;
+import com.Project_2_Location_Status_API.DTO.CovidStatsDTO;
 import com.Project_2_Location_Status_API.Services.CovidApiService;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,9 +9,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/covid-api")
@@ -24,14 +20,14 @@ public class CovidApiController {
     private CovidApiService covidApiService;
 
     @GetMapping("search") // covid-api/search?country=nameOfCountry
-    public ResponseEntity<CovidStats> getDataByCountry(@RequestParam String country){
+    public ResponseEntity<CovidStatsDTO> getDataByCountry(@RequestParam String country){
         return ResponseEntity.ok(covidApiService.getAllDataByCountry(country).getBody());
     }
 
     @GetMapping("calculate")
     public ResponseEntity<String> getStatusBasedOnLocation(@RequestParam String country) {
 
-        ResponseEntity<CovidStats> response = covidApiService.getAllDataByCountry(country);
+        ResponseEntity<CovidStatsDTO> response = covidApiService.getAllDataByCountry(country);
         Integer score = covidApiService.calculateStatus(response);
 
         if (score < 50) {
